+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 ## Unreleased
 
+## v1.5.3
+
+- [#266](https://github.com/chingor13/json_api_client/pull/266) - Fix default attributes being overridden
+- [#267](https://github.com/chingor13/json_api_client/pull/267) - Fix custom pagination params
+- [#271](https://github.com/chingor13/json_api_client/pull/271) - Fix assert_nil warnings
+- [#275](https://github.com/chingor13/json_api_client/pull/275) - Add error messages when destroy fails
+- [#277](https://github.com/chingor13/json_api_client/pull/277) - Correct handling of error source/pointer
+
+## v1.5.2
+
+- [#264](https://github.com/chingor13/json_api_client/pull/264) - Enable sparse fieldsets for nested models
+- [#263](https://github.com/chingor13/json_api_client/pull/263) - Fix initializing resource including relationships
+- [#260](https://github.com/chingor13/json_api_client/pull/260) - Use formatter for belongs_to keys
+
 ## v1.5.1
 
 - [#236](https://github.com/chingor13/json_api_client/pull/236) - Escape nested route keys